+++ app/lib/package/backend.dart
@@ -971,6 +971,7 @@ class PackageBackend {
     await purgePublisherCache(newPublisherId);
 
     if (email != null) {
+      await emailBackend.mirrorToSql(email!);
       await emailBackend.trySendOutgoingEmail(email!);
     }
     if (currentPublisherId != null) {
@@ -1574,6 +1575,7 @@ class PackageBackend {
     if (packagePublishedRecord != null) {
       await auditBackend.mirrorToSql(packagePublishedRecord!);
     }
+    await emailBackend.mirrorToSql(outgoingEmail);
     _logger.info('Upload successful. [package-uploaded]');
     _logger.info('Upload transaction completed in ${sw.elapsed}.');
     sw.reset();